+++ src/main/kotlin/com/fwcd/ktda/core/Position.kt
@@ -3,6 +3,6 @@ package com.fwcd.ktda.core
 /** A source code position. Line and column numbers are 1-indexed */
 class Position(
 	val source: Source,
-	val lineNumber: Int,
-	val columnNumber: Int? = null
+	val lineNumber: Long,
+	val columnNumber: Long? = null
 )
